broke files up by widget
