+++ packages/safety-suite-api/src/api/dashboardSink/util.ts
@@ -1,0 +1,7 @@
+import {Sort} from './types';
+
+export const convertSortsToStrings = (sorts: Sort[]): string[] => {
+  return sorts.map((sort) =>
+    sort.direction ? `${sort.column},${sort.direction}` : sort.column
+  );
+};
