+++ packages/safety-suite-api/src/api/index.ts
@@ -1,6 +1,7 @@
 export * from './authentication';
 export * from './configuration';
 export * from './customers';
+export * from './dashboardSink';
 export * from './documentLibrary';
 export * from './eForms';
 export * from './integrations';
